Moved IGCAPTgui from openstreetmap to gov.inl.igcapt.view
Lots of import fixes, and I had to make Dragtree7 public as well as IGCAPTgui.CurrentTypeUuidStr. To run I had to change the path for my own maven environment to the new location for IGCAPTgui. To do so, go to File->Project Repositories->Run and then change the path there.